laneData의 특수 데이터 값이 정상적으로 배당되지 않는 부분 수정

+++ src/Extension/Parser.hpp
@@ -39,7 +39,7 @@ typedef struct meta_data
 	string PLAYLEVEL;
 	string SONGID;
 	string WAVE;
-	float WAVEOFFSET;
+	double WAVEOFFSET;
 	string JACKET;
 	map<string, string> REQUEST;
 	int MEASUREBS;
